// calendar/src/domain/enums/months.enum.ts
export enum PersianMonth {
  JANUARY = "فروردین",
  FEBRUARY = "اردیبهشت",
  MARCH = "خرداد",
  APRIL = "تیر",
  MAY = "مرداد",
  JUNE = "شهریور",
  JULY = "مهر",
  AUGUST = "آبان",
  SEPTEMBER = "آذر",
  OCTOBER = "دی",
  NOVEMBER = "بهمن",
  DECEMBER = "اسفند",
}

export enum EnglishMonth {
  JANUARY = "January",
  FEBRUARY = "February",
  MARCH = "March",
  APRIL = "April",
  MAY = "May",
  JUNE = "June",
  JULY = "July",
  AUGUST = "August",
  SEPTEMBER = "September",
  OCTOBER = "October",
  NOVEMBER = "November",
  DECEMBER = "December",
}

export type Month = PersianMonth | EnglishMonth;